testccl/sqlccl: move a single file from sql package that needs CCL code
This commit breaks up `sql/run_control_test.go` file into two parts: one
that needs access to `ccl` code is moved into the new `ccl/testccl/sqlccl`
package whereas parts that don't need CCL code are kept in the same
place.

Note that there is another test (TestTelemetry) in `sql` package that
currently also needs access to CCL code.

Release note: None
