+++ pkg/BUILD.bazel
@@ -31,6 +31,8 @@ ALL_TESTS = [
     "//pkg/ccl/streamingccl/streamingutils:streamingutils_test",
     "//pkg/ccl/streamingccl/streamproducer:streamproducer_test",
     "//pkg/ccl/telemetryccl:telemetryccl_test",
+    "//pkg/ccl/testccl/sqlccl:sqlccl_test",
+    "//pkg/ccl/testccl/sqlccl:sqltestccl_test",
     "//pkg/ccl/testccl/workload/schemachange:schemachange_test",
     "//pkg/ccl/utilccl/sampledataccl:sampledataccl_test",
     "//pkg/ccl/utilccl:utilccl_test",
